+++ lib/src/cloud/cloud_error_reporters.dart
@@ -2,6 +2,7 @@
 
 export 'cloud_error_reporter_config.dart';
 export 'composite_error_reporter.dart';
+export 'console_error_reporter.dart';
 export 'firebase_crashlytics_reporter.dart';
 export 'http_error_reporter.dart';
 export 'sentry_error_reporter.dart';